+++ src/lock_box.c
@@ -1,47 +1,63 @@
 #include "crypto_box.h"
 
-void read_plaintext(FILE *input) {
+void lock_box(FILE *input, FILE *output) {
   size_t nread;
+  unsigned char mac_mac[crypto_onetimeauth_BYTES];
+  crypto_onetimeauth_state mac_mac_state;
+  crypto_onetimeauth_init(&mac_mac_state, key);
 
-  while(1) {
-    grow_ct(&ct, READ_BYTES);
+  if (isatty(fileno(output)))
+    fprintf(stderr, "WARNING: Writing ciphertext to terminal.\n");
 
-    nread = fread(&ct.data[ct.used], sizeof *ct.data, READ_BYTES, input);
+  init_ct(&ct);
+  while(!feof(input)) {
+    /* recycle ct */
+    ct.used = MAC_BYTES; /* reserve room for MAC */
+
+    /* read complete chunk, if possible */
+    grow_ct(&ct, CHUNK_BYTES);
+    nread = fread(&ct.data[ct.used], sizeof *ct.data, CHUNK_BYTES, input);
     ct.used += nread;
+    DEBUG_ONLY(hexDump("read plaintext chunk",
+          CT_AFTER_MAC(ct.data), PT_LEN(ct.used)));
 
-    if (READ_BYTES == nread) continue;
-    if (feof(input)) break;
-    if (ferror(input)) {
-      perror("Couldn't read plaintext");
+    if (nread < CHUNK_BYTES && ferror(input)) {
+      fprintf(stderr, "Couldn't read plaintext.\n");
       exit(EXIT_FAILURE);
     }
+
+    /* new nonce */
+    randombytes_buf(nonce, sizeof nonce);
+    DEBUG_ONLY(hexDump("nonce", nonce, sizeof nonce));
+
+    /* encrypt chunk */
+    crypto_secretbox_easy(ct.data, CT_AFTER_MAC(ct.data), PT_LEN(ct.used),
+        nonce, key);
+    DEBUG_ONLY(hexDump("chunk MAC", ct.data, MAC_BYTES));
+    DEBUG_ONLY(hexDump("ciphertext chunk", CT_AFTER_MAC(ct.data),
+          PT_LEN(ct.used)));
+
+    /* print ciphertext */
+    if (fwrite(nonce, sizeof nonce, 1, output) < 1 || /* nonce */
+        fwrite(ct.data, ct.used, 1, output) < 1) { /* MAC and CT */
+
+      perror("Couldn't write ciphertext");
+      exit(EXIT_FAILURE);
+    }
+
+    /* update mac_mac_state */
+    crypto_onetimeauth_update(&mac_mac_state, ct.data,
+        crypto_onetimeauth_BYTES);
   }
-  DEBUG_ONLY(hexDump("plaintext", CT_AFTER_MAC(ct.data), PT_LEN(ct.used)));
-}
+  free_ct(&ct);
 
-// in-place encryption + MAC
-void encrypt_then_mac(void) {
-  crypto_secretbox_easy(ct.data, CT_AFTER_MAC(ct.data), PT_LEN(ct.used), nonce, key);
-  DEBUG_ONLY(hexDump("MAC", ct.data, MAC_BYTES));
-  DEBUG_ONLY(hexDump("ciphertext", CT_AFTER_MAC(ct.data), PT_LEN(ct.used)));
-}
-
-void write_ciphertext(FILE *output) {
-  if (isatty(fileno(output))) {
-    fprintf(stderr, "WARNING: Writing ciphertext to terminal.\n");
-  }
-
-  fwrite(nonce, sizeof nonce, 1, output); // write nonce first
-  fwrite(ct.data, sizeof *ct.data, ct.used, output); // then MAC and CT
-  if (ferror(output)) {
+  /* print mac_mac */
+  crypto_onetimeauth_final(&mac_mac_state, mac_mac);
+  DEBUG_ONLY(hexDump("MAC of MACs", mac_mac, crypto_onetimeauth_BYTES));
+  if (fwrite(mac_mac, sizeof mac_mac, 1, output) < 1) {
     perror("Couldn't write ciphertext");
     exit(EXIT_FAILURE);
   }
-}
-
-void get_nonce(void) {
-  randombytes_buf(nonce, sizeof nonce);
-  DEBUG_ONLY(hexDump("nonce", nonce, sizeof nonce));
 }
 
 int main(int argc, char *argv[]) {
@@ -55,17 +71,10 @@ int main(int argc, char *argv[]) {
 
   key_mlock();
   get_key(&arguments, key);
-  get_nonce();
-
-  init_ct(&ct);
-  ct.used = MAC_BYTES; // reserve room for MAC
 
   FILE *input = open_input(&arguments);
-  read_plaintext(input);
+  lock_box(input, stdout);
   close_input(input);
-  encrypt_then_mac();
-  write_ciphertext(stdout);
-  free_ct(&ct);
 
   return EXIT_SUCCESS;
 }